Refactor remote providers with trait
